remove.sh: now control some errors

+++ myjenkins/remove.sh
@@ -1,6 +1,12 @@
 #!/bin/sh
-docker stop $(docker ps -aq --filter ancestor=myjenkins)
-docker rm -f $(docker ps -aq --filter ancestor=myjenkins)
-docker rmi -f $(docker images -aq --filter reference=myjenkins)
+if [ $(docker ps -aq --filter ancestor=myjenkins) ]; then
+    docker stop $(docker ps -aq --filter ancestor=myjenkins)
+    docker rm -f $(docker ps -aq --filter ancestor=myjenkins)
+fi
+if [ $(docker images -aq --filter reference=myjenkins) ]; then
+    docker rmi -f $(docker images -aq --filter reference=myjenkins)
+else
+    echo Already removed!
+fi
